Conflicts:
	nordismobile/src/com/panzyma/nm/view/ViewDevolucionEdit.java

+++ nordismobile/src/com/panzyma/nm/serviceproxy/DevolucionProductoLote.java
@@ -69,13 +69,13 @@ public class DevolucionProductoLote implements Parcelable
 	/**
 	 * @return the fechaVencimiento
 	 */
-	public int getFechaVencimiento() {
+	public long getFechaVencimiento() {
 		return fechaVencimiento;
 	}
 	/**
 	 * @param fechaVencimiento the fechaVencimiento to set
 	 */
-	public void setFechaVencimiento(int fechaVencimiento) {
+	public void setFechaVencimiento(long fechaVencimiento) {
 		this.fechaVencimiento = fechaVencimiento;
 	}
 	/**
@@ -152,7 +152,7 @@ public class DevolucionProductoLote implements Parcelable
 	private long id;
 	private long objLoteID;
 	private String numeroLote;
-	private int fechaVencimiento;
+	private long fechaVencimiento;
 	private int cantidadDevuelta;
 	private boolean fueraPolitica;
 	private int cantidadDespachada;
@@ -180,7 +180,7 @@ public class DevolucionProductoLote implements Parcelable
 		parcel.writeLong(  id );
 		parcel.writeLong(  objLoteID );
 		parcel.writeString( numeroLote );
-		parcel.writeInt(  fechaVencimiento );
+		parcel.writeLong(  fechaVencimiento );
 		parcel.writeInt( cantidadDevuelta );
 		parcel.writeInt( fueraPolitica ? 1 : 0 );
 		parcel.writeInt( cantidadDespachada );
